Fix compiler warning (-Wchar-subscript) (#1259)
ccstruct/seam.cpp:66:26: warning:
 array subscript has type 'char' [-Wchar-subscripts]

Fix it by using an unsigned index and use the same type for related values.

+++ ccstruct/seam.cpp
@@ -62,7 +62,7 @@ void SEAM::CombineWith(const SEAM& other) {
   location_ += other.location_;
   location_ /= 2;
 
-  for (int s = 0; s < other.num_splits_ && num_splits_ < kMaxNumSplits; ++s)
+  for (uint8_t s = 0; s < other.num_splits_ && num_splits_ < kMaxNumSplits; ++s)
     splits_[num_splits_++] = other.splits_[s];
 }
